Get rid of percy/capybara structure and Percy::Capybara module.
