--- lib/percy/capybara/version.rb
@@ -1,5 +1,0 @@
-module Percy
-  module Capybara
-    VERSION = '4.0.0'.freeze
-  end
-end
