+++ lib/version.rb
@@ -1,0 +1,3 @@
+module Percy
+  VERSION = '4.0.0'.freeze
+end
